LayoutTests/platform/blackberry/accessibility/image-map1-expected.txt

 
This tests that you can reach the links within an image map.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL body.childAtIndex(0).role should be AXRole: AXLink. Was .
FAIL body.childAtIndex(0).title should be AXTitle: Link1. Was .
FAIL body.childAtIndex(1).role should be AXRole: AXLink. Was .
FAIL body.childAtIndex(1).title should be AXTitle: Link2. Was .
FAIL body.childAtIndex(2).role should be AXRole: AXLink. Was .
FAIL body.childAtIndex(2).title should be AXTitle: Link3. Was .
PASS successfullyParsed is true

TEST COMPLETE
